Abstract Wagtail mocks into separate class
