+++ exams/views_test.py
@@ -1,42 +1,13 @@
 """
 Tests for exam views
 """
-from unittest.mock import patch
-from django.test import SimpleTestCase
-
-# These fake middlewares allow us to use Django.test.SimpleTestCase,
-# which runs faster and doesn't touch the database at all.
-
-
-class FakeSiteMiddleware(object):  # pylint: disable=missing-docstring
-    def process_request(self, request):  # pylint: disable=missing-docstring,no-self-use
-        request.site = None
-        return None
-
-
-class FakeRedirectMiddleware(object):  # pylint: disable=missing-docstring
-    def process_request(self, request):  # pylint: disable=missing-docstring,no-self-use,unused-argument
-        return None
+from micromasters.test import SimpleTestCase
 
 
 class PearsonSSOCallbackTests(SimpleTestCase):
     """
     Tests for Pearson callback URLs
     """
-    def setUp(self):
-        site_patcher = patch(
-            'wagtail.wagtailcore.middleware.SiteMiddleware',
-            new=FakeSiteMiddleware
-        )
-        redirect_patcher = patch(
-            'wagtail.wagtailredirects.middleware.RedirectMiddleware',
-            new=FakeRedirectMiddleware
-        )
-        site_patcher.start()
-        self.addCleanup(site_patcher.stop)
-        redirect_patcher.start()
-        self.addCleanup(redirect_patcher.stop)
-
     def test_success(self):
         """
         Test /pearson/success URL
